requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-multipart
surya-ocr
Pillow
